+++ src/main/java/com/codeborne/selenide/SelenideElement.java
@@ -741,7 +741,7 @@ public interface SelenideElement extends WebElement, WrapsDriver, WrapsElement, 
   @Override void click();
 
   /**
-   * Click the element with a relative offset from the upper left corner of the element
+   * Click the element with a relative offset from the center of the element
    */
   void click(int offsetX, int offsetY);