+++ src/main/java/com/codeborne/selenide/conditions/SelectedText.java
@@ -28,6 +28,6 @@ public class SelectedText extends Condition {
 
   @Override
   public String toString() {
-    return getName() + " '" + expectedText + '\'';
+    return String.format("%s '%s'", getName(), expectedText);
   }
 }