+++ src/main/java/com/codeborne/selenide/ex/TextsSizeMismatch.java
@@ -1,0 +1,24 @@
+package com.codeborne.selenide.ex;
+
+import com.codeborne.selenide.impl.WebElementsCollection;
+
+import java.util.List;
+
+import static java.lang.System.lineSeparator;
+
+public class TextsSizeMismatch extends UIAssertionError {
+  public TextsSizeMismatch(WebElementsCollection collection, List<String> actualTexts,
+                       List<String> expectedTexts, String explanation, long timeoutMs) {
+    super(collection.driver(),
+        lineSeparator() + "Actual: " + actualTexts + ", List size: " + actualTexts.size() +
+        lineSeparator() + "Expected: " + expectedTexts + ", List size: " + expectedTexts.size() +
+        (explanation == null ? "" : lineSeparator() + "Because: " + explanation) +
+        lineSeparator() + "Collection: " + collection.description());
+    super.timeoutMs = timeoutMs;
+  }
+
+  @Override
+  public String toString() {
+    return getClass().getSimpleName() + ' ' + getMessage() + uiDetails();
+  }
+}