+++ statics/src/main/java/com/codeborne/selenide/junit5/ScreenShooterExtension.java
@@ -19,6 +19,51 @@ import static com.codeborne.selenide.ex.ErrorMessages.screenshot;
 import static com.codeborne.selenide.logevents.LogEvent.EventStatus.FAIL;
 
 /**
+ * Use this class to automatically take screenshots in case of ANY errors in tests (not only Selenide errors).
+ *
+ * How to use in Java:
+ * <pre>
+ * {@code
+ *    @ExtendWith({ScreenShooterExtension.class})
+ *    public class MyTest {...}
+ * }
+ * </pre>
+ *
+ * How to use in Java (with customization):
+ * <pre>
+ * {@code
+ *   public class MyTest {
+ *     @RegisterExtension
+ *     static ScreenShooterExtension screenshotEmAll = new ScreenShooterExtension(true);
+ *     ...
+ *   }
+ * }
+ * </pre>
+ *
+ * How to use in Kotlin:
+ *
+ * <pre>
+ *   {@code
+ *     @ExtendWith(ScreenShooterExtension::class)
+ *     public class MyTest {...}
+ *   }
+ * </pre>
+ *
+ * How to use in Kotlin (with customization):
+ *
+ * <pre>
+ * {@code
+ *   public class MyTest {
+ *     companion object {
+ *       @JvmField
+ *       @RegisterExtension
+ *       val screenshotEmAll: ScreenShooterExtension = ScreenShooterExtension(true);
+ *     }
+ *     ...
+ *   }
+ * }
+ * </pre>
+ *
  * @author Aliaksandr Rasolka
  * @since 4.12.2
  */
@@ -96,6 +141,4 @@ public class ScreenShooterExtension implements AfterAllCallback, BeforeAllCallba
   protected String getScreenshot() {
     return screenshot(driver());
   }
-
-
 }